Promisify request module
Using the presage promise utility library, the request module has been
promisified. This removes the last callback-based logic in the library
to be abstracted away, and the new Promise returned by request-promise
can be returned directly by the perform-request function.

Includes a series of unit tests for request-promise that make certain
that various errors are thrown/rejected as expected should it be
provided with improper arguments.