Add unit testing for client.scroll
client.scroll was the only client method not subject to unit tests.
During testing, it was revealed that perform-request was improperly
processing headers. The logic will now attempt to use the lowercase
header names first, and then the Initial Case header names if the
lowercase ones are not present in the response object.

This brings the project to 100% unit test coverage.